# contracts/openapi/weave-admin-openapi.json
{
  "components": {
...
    "schemas": {
...
      "ApiErrorResponse": {
        "description": "Stable JSON error envelope returned by protected API endpoints.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "example": "unauthorized",
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "description": "Non-secret diagnostic details for this failure.",
            "type": "object"
          },
          "memberImpact": {
            "description": "Optional member-facing impact state or recovery hint owned by the backend facade.",
            "example": "unavailable",
            "type": "string"
          },
          "message": {
            "description": "Support-safe explanation of what failed.",
            "example": "Bearer authentication is required to access this endpoint.",
            "type": "string"
          },
          "requestId": {
            "description": "Correlation identifier for support and log lookup.",
            "example": "8c2f4d7a-4d5f-4f88-92e8-4fcbb81dd527",
            "type": "string"
          },
          "supportRef": {
            "description": "Stable reference members can share with support without provider payloads.",
            "example": "support:8c2f4d7a-4d5f-4f88-92e8-4fcbb81dd527",
            "type": "string"
          }
        },
        "required": [
          "code",
          "details",
          "message",
          "requestId",
          "supportRef"
        ],
        "type": "object"
      },
...
      "MemberLifecycleOperationResponse": {
        "properties": {
          "memberHandle": {
            "type": "string"
          },
          "outcome": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "MemberOffboardingRequest": {
        "properties": {
          "confirmation": {
            "pattern": "OFFBOARD_MEMBER",
            "type": "string"
          }
        },
        "required": [
          "confirmation"
        ],
        "type": "object"
      },
...
      "OrganizationMemberPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrganizationMemberResponse"
            },
            "type": "array"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "OrganizationMemberResponse": {
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "memberHandle": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "capabilities",
          "enabled"
        ],
        "type": "object"
      },
      "OrganizationMemberUpdateRequest": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "role": {
            "minLength": 1,
            "pattern": "owner|admin|member|guest",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "type": "object"
      },
...
      "SpaceMemberChangeRequest": {
        "properties": {
          "permissionLevel": {
            "description": "Hierarchical current grant: VIEW, VIEW+EDIT, or VIEW+EDIT+ADMIN.",
            "enum": [
              "VIEW",
              "EDIT",
              "ADMIN"
            ],
            "type": "string"
          }
        },
        "required": [
          "permissionLevel"
        ],
        "type": "object"
      },
...
      "SpaceMemberResponse": {
        "properties": {
          "accountRef": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "VIEW",
                "EDIT",
                "ADMIN"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "accountRef",
          "permissions"
        ],
        "type": "object"
      },
...
      "WeaverEntitlementUpdateRequest": {
        "properties": {
          "entitled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "owner-bootstrap-token": {
        "description": "Rotatable SecretRef credential for the empty-realm owner invitation only.",
        "in": "header",
        "name": "X-Weave-Bootstrap-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "JWT-protected product API for workspace capabilities and orchestration.",
    "title": "Weave Backend API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/organizations/{organizationId}/members": {
      "get": {
        "operationId": "listOrganizationMembers",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberPageResponse"
                }
              }
            },
            "description": "Paged organization members visible to the administrator."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      }
    },
    "/api/admin/organizations/{organizationId}/members/{memberHandle}": {
      "get": {
        "operationId": "getOrganizationMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberHandle",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberResponse"
                }
              }
            },
            "description": "Current organization member and version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      },
      "patch": {
        "operationId": "updateOrganizationMemberAccess",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberHandle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationMemberUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberResponse"
                }
              }
            },
            "description": "Updated organization member and version."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      }
    },
    "/api/admin/organizations/{organizationId}/members/{memberHandle}/capabilities/weaver": {
      "put": {
        "operationId": "updateOrganizationMemberWeaverEntitlement",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberHandle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeaverEntitlementUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberResponse"
                }
              }
            },
            "description": "Organization member after Weaver entitlement update."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      }
    },
    "/api/admin/organizations/{organizationId}/members/{memberHandle}/offboarding": {
      "post": {
        "operationId": "offboardOrganizationMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberHandle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberOffboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberLifecycleOperationResponse"
                }
              }
            },
            "description": "Support-safe member offboarding result."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      }
    },
    "/api/admin/organizations/{organizationId}/members/{memberHandle}/session-revocations": {
      "post": {
        "operationId": "revokeOrganizationMemberSessions",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberHandle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberLifecycleOperationResponse"
                }
              }
            },
            "description": "Support-safe session revocation result."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The caller cannot administer the configured organization."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "tags": [
          "Organization members"
        ]
      }
    },
...
    "/api/admin/spaces/{spaceRef}/members/{accountRef}": {
      "delete": {
        "operationId": "deleteSpaceMember",
        "parameters": [
          {
            "in": "path",
            "name": "spaceRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "accountRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Current grant revoked."
          },
          "400": {
            "description": "Invalid member precondition."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Owner/admin policy or organization denied."
          },
          "404": {
            "description": "Member unavailable."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Existing Space differs from setup intent."
          },
          "412": {
            "description": "Member precondition is stale."
          },
          "428": {
            "description": "Member precondition required."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "summary": "Revoke current Space membership",
        "tags": [
          "Spaces Admin"
        ]
      },
      "get": {
        "operationId": "getSpaceMember",
        "parameters": [
          {
            "in": "path",
            "name": "spaceRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "accountRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Current member grant.",
            "headers": {
              "ETag": {
                "description": "Strong current member version.",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Owner/admin policy or organization denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Current member unavailable."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Existing Space differs from setup intent."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revoked grant; ETag identifies the tombstone for explicit regrant.",
            "headers": {
              "ETag": {
                "description": "Strong revoked member version.",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "summary": "Read current Space member grant and version",
        "tags": [
          "Spaces Admin"
        ]
      },
      "put": {
        "operationId": "putSpaceMember",
        "parameters": [
          {
            "in": "path",
            "name": "spaceRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "accountRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpaceMemberChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Versioned current member grant.",
            "headers": {
              "ETag": {
                "description": "Strong current member version required for replacement or revocation.",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Invalid grant or precondition."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Owner/admin policy or organization denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Member unavailable for replacement."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Existing Space differs from setup intent."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Member precondition is stale."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceMemberResponse"
                }
              }
            },
            "description": "Member precondition required."
          }
        },
        "security": [
          {
            "bearer-jwt": []
          }
        ],
        "summary": "Grant or replace current Space membership",
        "tags": [
          "Spaces Admin"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}